Refactor command handling into module
